refactor : seperated 'CategorySetup' and 'ArticleSetupModal' - reason : ArticleSetupModal has various setup list. so it seems to seperate each other.
